requirements.txt
# 基础依赖
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0
pyyaml>=5.4.0

# Qlib - 量化投资平台
qlib>=0.9.0

# 机器学习
lightgbm>=3.3.0
scikit-learn>=1.0.0
xgboost>=1.5.0  # 可选
catboost>=1.0.0  # 可选

# vn.py - 交易平台
vnpy>=3.0.0
vnpy-ctp>=3.0.0  # CTP接口
vnpy-ctastrategy>=3.0.0  # CTA策略

# 可视化
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.0.0

# 工具
tqdm>=4.62.0
joblib>=1.1.0

# Jupyter支持（可选）
jupyter>=1.0.0
ipykernel>=6.0.0

# 数据库（可选）
pymongo>=3.12.0
sqlalchemy>=1.4.0

# GPU加速（可选，需要CUDA环境）
# 如果使用GPU版本的LightGBM，请安装:
# pip install lightgbm --install-option=--gpu

